move Player to its own folder
